Addressing Michael's comments
